watchmate/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
click==8.0.4
distlib==0.3.4
Django==4.0.3
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.4.2
joblib==1.1.0
nltk==3.7
pipenv==2021.11.23
platformdirs==2.4.1
PyJWT==2.6.0
PyMySQL==1.0.2
pytz==2022.7
regex==2022.3.15
six==1.16.0
sqlparse==0.4.2
tqdm==4.63.0
virtualenv==20.13.0
virtualenv-clone==0.5.7
